% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dfEff}
\alias{dfEff}
\alias{dfEff.multModel}
\title{Effective degrees of freedom.}
\usage{
dfEff(model, par, ...)

\method{dfEff}{multModel}(model, par = model$B, method = model$dfMethod,
  ...)
}
\arguments{
\item{model}{the model for which the degrees of freedom is to be computed.}

\item{par}{a \code{numeric}. The parameter vector.}

\item{method}{a \code{character}. The method for estimation of degrees of freedom. The default is to use
the method specified by the model. The possible values are \code{'pen'}, \code{'sub'},
\code{'con'} and \code{'nonzero'}.}

\item{...}{}
}
\value{
a numeric.
}
\description{
Estimation of the effective degrees of freedom.
}
\author{
Niels Richard Hansen \email{Niels.R.Hansen@math.ku.dk}
}
\seealso{
\code{\link{risk}}, \code{\link{riskHat}}
}

