\name{dfEffec}
\alias{dfEffec}
\title{Computation of effective degrees of freedom.}
\usage{
dfEffec(object, par, ...)
}
\arguments{
  \item{object}{for which the degrees of freedom is to be
  computed.}

  \item{par}{the parameter vector.}

  \item{...}{}
}
\value{
a numeric.
}
\description{
Computation of effective degrees of freedom.
}
\author{
Niels Richard Hansen \email{Niels.R.Hansen@math.ku.dk}
}
\seealso{
\code{\link{risk}}, \code{\link{riskHat}}
}

