\name{dpredict}
\alias{dpredict}
\title{Derivative of the predictor}
\usage{
dpredict(object, par, ...)
}
\arguments{
  \item{object}{an object of class \code{multModel}.}

  \item{par}{the parameter vector.}
}
\description{
Derivative of the predictor
}
\author{
Niels Richard Hansen \email{Niels.R.Hansen@math.ku.dk}
}
\seealso{
\code{\link{predict.multModel}}, \code{\link{multModel}}
}

