% Generated by roxygen2 (4.0.1): do not edit by hand
\name{grad}
\alias{grad}
\alias{grad.multModel}
\title{Computation of the gradient of the loss function}
\usage{
grad(model, par, ...)

\method{grad}{multModel}(model, par = model$B, ...)
}
\arguments{
\item{model}{for which the gradient of the loss is to be computed.}

\item{par}{the parameter vector.}

\item{...}{}
}
\value{
a numeric vector.
}
\description{
Computation of the gradient of the loss function
}
\author{
Niels Richard Hansen \email{Niels.R.Hansen@math.ku.dk}
}
\seealso{
\code{\link{loss}}
}

