% Generated by roxygen2 (4.0.1): do not edit by hand
\name{multModel}
\alias{multModel}
\title{Multivariate process models constructor}
\usage{
multModel(y, x, t, suff = TRUE, weights = "unit", sigma = 1)
}
\arguments{
\item{y}{a \code{numeric}. A response matrix of dimensions d by m.}

\item{x}{a \code{numeric}. A predictor matrix of dimensions d by m.}

\item{t}{a \code{numeric}. A single time step size, or a vector of time step sizes of length m.}

\item{suff}{a \code{logical}. Should sufficient statistics and the MLE be computed.}

\item{weights}{a \code{character} or \code{numeric}. Specifies parameter weights. Default value
\code{'unit'} means that all parameters are weighted equally with a unit weight,
whereas \code{'adaptive'} means that the parameters are weighted inverse proportionally
to the MLE (will only be used if \code{suff = TRUE}). Alternatively, a numeric specifying
the individual parameter weights.}
}
\value{
an object of class \code{multModel} containing the elements
 \tabular{ll}{
 y \tab the d by m data matrix \cr
 x \tab the d by m data matrix \cr
 t \tab the time step(s) \cr
 ss \tab sum of squares (computed if \code{suff = TRUE}) \cr
 xyt \tab d by d crossproduct of x and y (computed if \code{suff = TRUE}) \cr
 xxt \tab d by d crossproduct of x (computed if \code{suff = TRUE}) \cr
 B \tab the MLE if \code{suff = TRUE}, and the zero matrix otherwise \cr
 w \tab a matrix of parameter weights \cr
 d \tab dimension of the process \cr
 m \tab number of d-dimensional observations \cr
 n \tab equals \code{m * d} \cr
 p \tab quals \code{d * d} \cr
 sigma \tab the observation variance parameter \cr
 status \tab 0 means no problems or errors encountered \cr
 msg \tab error messages or other notable conditions
 }
}
\description{
Constructs multivariate models for dynamic processes.
}
\author{
Niels Richard Hansen \email{Niels.R.Hansen@math.ku.dk}
}

