\name{predict.multModel}
\alias{predict.multModel}
\title{Prediction for multivariate models}
\usage{
\method{predict}{multModel}(object, par, ...)
}
\arguments{
  \item{object}{an object of class \code{multModel}.}

  \item{par}{the parameter vector.}
}
\description{
Prediction for multivariate models
}
\author{
Niels Richard Hansen \email{Niels.R.Hansen@math.ku.dk}
}
\seealso{
\code{\link{predict}}, \code{\link{multModel}}
}

