% Generated by roxygen2 (4.0.1): do not edit by hand
\name{response}
\alias{response}
\alias{response.multModel}
\title{Returns the response}
\usage{
response(model, ...)

\method{response}{multModel}(model, ...)
}
\arguments{
\item{object}{an object from which the response is to be obtained.}
}
\description{
Returns the response
}
\author{
Niels Richard Hansen \email{Niels.R.Hansen@math.ku.dk}
}
\seealso{
\code{\link{multModel}}
}

