\name{risk}
\alias{risk}
\title{Computation of the squared error risk.}
\usage{
risk(model, par0, par, ...)
}
\arguments{
  \item{object}{for which the risk is to be computed.}

  \item{par0}{the best / true parameter vector.}

  \item{par}{the estimated parameter vector.}

  \item{...}{}
}
\value{
a numeric.
}
\description{
Computation of the squared error risk.
}
\author{
Niels Richard Hansen \email{Niels.R.Hansen@math.ku.dk}
}
\seealso{
\code{\link{dfEff}}, \code{\link{riskHat}}
}

