\docType{package}
\name{smde}
\alias{smde}
\alias{smde-package}
\title{Sparse Multivariate Differential Equations}
\description{
Estimation of coefficients in multivariate ordinary or
stochastic differential equations using penalized nonlinear
least squares methods.
}
\details{
\tabular{ll}{ Package: \tab smde \cr Type: \tab Package\cr
Version: \tab 0.3\cr Date: \tab 2014-01-17\cr License: \tab
GPL (>= 2)\cr LazyLoad: \tab yes } Estimation of high
dimensional, discretely observed diffusions using a
strategy which yields a sparse representation of the drift
term.
}
\author{
Niels Richard Hansen \email{Niels.R.Hansen
}
\keyword{package}

