% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stepOptim}
\alias{stepOptim}
\title{Stepwise optimization}
\usage{
stepOptim(f, gr, par, direction = "backward", scope, ...)
}
\arguments{
\item{f}{a (loss) function to minimize}

\item{gr}{the gradient of f.}

\item{par}{the initial parameter vector.}

\item{direction}{the mode of the stepwise search. The default is \code{'backward'}.
The alternative is \code{'forward'}.}

\item{scope}{either a sequence of indices indicating the nonzero entries in the largest model,
or a list containing the components \code{lower} and \code{upper}, each being a sequence
of indices indicating the nonzero entries in the smallest and the largest model, respectively.}
}
\value{
A \code{list} of length 4. The first element, \code{lambda}, is the sequence of dimensions,
        and the second, \code{beta}, is the matrix of parameter estimates. Each column in \code{beta}
        corresponds to an entry in \code{lambda}. The third element is \code{status}, where 0 means
        convergence, and 4 indicates convergence problems. The last element \code{msg} gives details
        on convergence status.
}
\description{
Stepwise estimation for a (sub)model specified by a scope of nonzero parameters.
}
\details{
The function implements a generic stepwise optimization algorithm. The optimization is done with \code{optim} using
the BFGS algorithm. The function searches either forward or backwards in the parameter vector, and in
each step it chooses among the possible models the best fitting model as measured by the loss function.
}
\author{
Niels Richard Hansen \email{Niels.R.Hansen@math.ku.dk}
}

