% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dpredict}
\alias{dpredict}
\alias{dpredict.multModel}
\title{Derivative of the predictor}
\usage{
dpredict(object, par, ...)

\method{dpredict}{multModel}(object, par, ...)
}
\arguments{
\item{object}{an object of class \code{multModel}.}

\item{par}{the parameter vector.}
}
\value{
An n by p matrix.
}
\description{
Derivative of the predictor
}
\author{
Niels Richard Hansen \email{Niels.R.Hansen@math.ku.dk}
}
\seealso{
\code{\link{predict.multModel}}, \code{\link{multModel}}
}

