% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fit}
\alias{fit}
\alias{fit.multModel}
\title{Fitting a multivariate process model}
\usage{
fit(model, ...)

\method{fit}{multModel}(model, method = "cd", scope = seq(1, model$p), ...)
}
\arguments{
\item{model}{}

\item{method}{a \code{character}. The default, \code{"cd"}, means a standard coordinate descent algorithm.
while \code{'mf'} means the same coordinate descent algorithm, but using a matrix free
implementation. A choice of \code{'forward'} or \code{'backward'} means least squares model search.}

\item{scope}{a \code{numeric} or a \code{list}. See \code{\link{stepOptim}}.}

\item{...}{additional arguments passed on to the optimization algorithm.}
}
\value{
a \code{multModel} object with the additional elements
 \tabular{ll}{
 lambda \tab the sequence of tuning parameters \cr
 Blambda \tab the corresponding matrix of estimated parameters. The i'th column corresponds to the i'th tuning parameter. \cr
 status \tab either 0 (meaning no errors), 1 (convergence), 2 (MLE not computed) or 3 (other errors)
              indicating different problems or errors \cr
 msg \tab errors encountered or other notable conditions
 }
}
\description{
Estimation of parameters in a multivariate process model. The function fits the models by minimizing
an l1-penalized squared error loss by a coordinate wise descent algorithm. Alternatively, the function
can do unpenalized least squares forward or backward model search.
}
\details{
The stepwise model search algorithms as well as the algorithm for l1-panalized estimation return
a sequence of models parametrized by a vector of tuning parameters. For the model search, the tuning
parameter is the dimension of the model (the number of nonzero entries in the parameter vector). For
the l1-panalized models, the tuning parameter is a sequence of penalty parameters.

The scope argument specifies the smallest and largest model fitted. They must be nested. The zero weights
of the \code{multModel} object are used to specify the smallest model, if the scope argument does not
contain a smallest model.
}
\seealso{
\code{\link{multModel}}, \code{\link{coordinateDescent}}, \code{\link{coordinateDescentMF}},
        \code{\link{stepOptim}}.
}

