\name{grad}
\alias{grad}
\title{Computation of the gradient of the loss function}
\usage{
grad(model, par, ...)
}
\arguments{
  \item{model}{for which the gradient of the loss is to be
  computed.}

  \item{par}{the parameter vector.}

  \item{...}{}
}
\value{
a numeric vector.
}
\description{
Computation of the gradient of the loss function
}
\author{
Niels Richard Hansen \email{Niels.R.Hansen@math.ku.dk}
}
\seealso{
\code{\link{loss}}
}

