% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loss}
\alias{loss}
\alias{loss.multModel}
\title{Computation of the squared error loss function}
\usage{
loss(model, par, ...)

\method{loss}{multModel}(model, par = model$B, ...)
}
\arguments{
\item{object}{for which the loss is to be computed.}

\item{par}{the parameter vector.}
}
\value{
a numeric.
}
\description{
Computation of the squared error loss function
}
\author{
Niels Richard Hansen \email{Niels.R.Hansen@math.ku.dk}
}
\seealso{
\code{\link{grad}}
}

