\name{response}
\alias{response}
\title{Returns the response}
\usage{
response(model, ...)
}
\arguments{
  \item{object}{an object from which the response is to be
  obtained.}
}
\description{
Returns the response
}
\author{
Niels Richard Hansen \email{Niels.R.Hansen@math.ku.dk}
}
\seealso{
\code{\link{multModel}}
}

