% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{smde}
\alias{smde}
\alias{smde-package}
\title{Sparse Multivariate Differential Equations}
\description{
Estimation of coefficients in multivariate ordinary or stochastic
differential equations using l1-penalized nonlinear least squares methods.
}
\details{
\tabular{ll}{
Package: \tab smde \cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2014-01-17\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes
}

Estimation of coefficients in multivariate and high dimensional models of dynamic
 systems modeled using ordinary or stochastic differential equations. The methods
 are based on l1-penalized nonlinear least squares to yield sparse representations.
}
\author{
Niels Richard Hansen \email{Niels.R.Hansen@math.ku.dk}
}

